requirements.txt
aiogram==3.1.1
python-dotenv==1.0.0
g4f==0.1.7.9
SQLAlchemy==2.0.22
alembic==1.12.1
python-decouple==3.8
